requirements.txt
BeautifulSoup
django-pagination==1.0.5